Fixing a crash exposed by the kn0ck0ut LV2 plugin.
After reporting the crash to the kn0ck0ut LV2 maintainer, he pointed out that
Audacity wasn't connecting to all required LV2 ports.  In this case it was the
latency port.  This corrects that oversight.

In addition, since I was fiddling with the latency port, I figured I might as
well go ahead and add the necessary code to compensate for it.




git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13139 47890b92-0858-11df-a26f-8b716316a5bc