Reset the selection time after a preview.
Steve found that if you preview an effect and, say, cancel it
before it finishes preparing, an apply will not modify all of
the originally selected audio...just up to the preview length.

This is very easy to reproduce using the Paulstretch on a 30
second clip...just click stop on the "Preparing preview"
dialog and the click apply.



git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13405 47890b92-0858-11df-a26f-8b716316a5bc