Fix the bad rendering of text on OSX
The problem was because I added double buffering to reduce
flicker on Windows.  But, OSX already does its own buffering
and adding more causes text to be rendered incorrectly.

This happened a long time ago on the track panel and when
Bill gave the screenshot, my aging brain slowly remembered
the cause.



git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13413 47890b92-0858-11df-a26f-8b716316a5bc