Now have three varieties of the Meter Toolbar.
We have separate record and play meters.  The original kind of meter is now called a combined meter.  I've kept it because it can be useful when undocked if you do want both meters.  I've also fixed it so that if made very narrow the meters stack vertically just as they already did horizontally.

git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13480 47890b92-0858-11df-a26f-8b716316a5bc