2nd attempt to Add the ability to build without trashing the source tree
In r13510 I had neglected to build with all local libraries and some of
them needed attention.

So, I also took the opportunity to work out the locale directory and
how to keep it unmolested as well.  As a result, all locales are rebuilt
as expected, but into the "build" directory.

As a bonus you may now test Audacity from the "build" directory and have
Nyquist plugins and message catalogs available (so you can test other 
languages again without havint to install).

So, again:

mkdir buildme
cd buildme
../configure
make
./audacity




git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13513 47890b92-0858-11df-a26f-8b716316a5bc