Realtime preview round 2
This changes the realtime preview from using 1 effect for all tracks
to an effect per track (logical track).  This should clear up the 
bad audio when more than one track (or a stereo track) is present.

An unfortunate side effect is that meter effects no longer work since
the one presented to the user is not the one doing the actual work.
Suggestions on how to remedy this are welcome.



git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13517 47890b92-0858-11df-a26f-8b716316a5bc