Provide a much easier way of building wxWidgets
I've added another project to the Audacity solution called
wxWidgets-2.8.12.  And updated the compile.txt file with
instructions on how to build everything.

It automatically copies our setup.h and the fixed window.h
to the proper locations in the wx build tree.

It uses NMAKE to build wxWidgets using the makefile in the
wx tree.  (No, I did not try to duplicate the wx project
files. ;-))

Give it a try.  I think you'll be surprised how easy it will
be for new users.



git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13540 47890b92-0858-11df-a26f-8b716316a5bc