Spectral selection cursor and mode selection changes from Paul Licameli.
This also now shows a new green 'play cursor' when ctrl is held down on wave track, and shift being held down is signalled by a cursor change to finger-pointer too.

I also accidentally enabled EXPERIMENTAL_MIDI_OUT.  We may disable that in December for release.  Leaving it in for now.

git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13555 47890b92-0858-11df-a26f-8b716316a5bc