Fix type problem with wxWidgets 3.0: Use enum wxRasterOperationMode instead of int.  [Patch 0022] from martin@steghoefer.eu
This change from the old int type to the real enum wxRasterOperationMode was missed during the fix of similar problems in r13403 (when similar fixes were applied), probably because ImageRoll and the affected variable are currently unused and the problem therefore doesn't cause a compilation error and is only fixed preventively.

Compatibility with wxWidgets 2.8 is assured thanks to a conditional definition of wxRasterOperationMode as int in ImageRoll.h, introduced in r13403.

git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13605 47890b92-0858-11df-a26f-8b716316a5bc