Only allow one realtime effect at a time
Warning...this required a small hack that allowed updating the
menus even if the toplevel frame is not the project window.  This
is because with the active frame actually becomes the effect dialog
and UpdateMenus() bailed if it wasn't the project frame.

This should be removed when we get multiple effect support.



git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13667 47890b92-0858-11df-a26f-8b716316a5bc