Resolves two issues
On Windows, when the button access keys were used, the focus
would jump to the associated button.  This would be a real
pain for keyboard user since they would then have to TAB back
to the control they were on to make further adjustments.

By using the accelerator table instead of the normal mnemonic,
the access key isn't directly tied to the button, so focus
remains untouched.

On GTK, this actually makes the accelerators work.  It seems
that normal mnemonics do not work with a wxBitmapButton, so
using the accelerator table provides the access keys.



git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13728 47890b92-0858-11df-a26f-8b716316a5bc