Reinstate close handle for AudioUnits
Closing the window via the frame's close button was
not being recognized by wxWidgets.  I'm sure it's
because of the of the additional event handler, but
the close event should have been sent on to the wx
handler since it wasn't handled.

No matter...just capture the event and close the
window manually.



git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13731 47890b92-0858-11df-a26f-8b716316a5bc