Fix for bug #40.
Whenever a menu is NOT open, the accelerator keys are
stripped from the menu items so that the O/S doesn't even
know that accelerators are being used (we do it internally).

While this stripping was being down when the menus are set
up, it was NOT being down when the Undo and Redo menu items
were modified to include the "last action".

So, the fix is to strip the accelerator then as well.



git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13739 47890b92-0858-11df-a26f-8b716316a5bc