Fix the weird mouse pointer behavior over toolbar resizers
I couldn't handle it anymore.  The darn pointer would seemingly
switch to left/right arrows whenever it felt like it...and stay
that way.

Actually, it was when it passed over the toolbar resizer when
docked.  The problem was that it wouldn't change back to a normal
pointer because it didn't have the events it needed to do that.

So, I moved the resizer logic into it's own window and now the
pointer changes like it should.

As a bonus, we get a tooltip so the user will know what to do
when the pointer changes as it passes over the resizer.



git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13741 47890b92-0858-11df-a26f-8b716316a5bc