More pointer madness fixes
This round fixes Mac pointers when mouse transitions from
the main project window to an open effect window in GUI
mode and back.

Not all of the cases are fixed with VST overlay windows
being the final holdout.  Couldn't figure out the cause
so will just deal with it when we convert to wx3.

(I GET TO REDO ALL OF THIS FOR WX3...WOOHOO!!!)



git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13762 47890b92-0858-11df-a26f-8b716316a5bc