Cleanup locale makefiles
This will remove that pesky Makefile.in that would get in
the way when trying to commit.  It doesn't need to be in 
SVN since it is regenerated each build.

And this will restore Makefile.in.in to the unmodified
version there was a risk of losing those changes when/if
autopoint was rerun.

The changes were moved to configure.ac.



git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13837 47890b92-0858-11df-a26f-8b716316a5bc