Fix for problem caused by r5527, circa Sep. 2006
The problem was that if the user chose to quit Audacity when
the auto recovery dialog was displayed, Audacity would crash
because the command line was still parsed and any filenames
listed on the command line would still be opened.  But, that
would failed because "quit" processing had already been done.

Anyway, a simple one line fix.



git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13857 47890b92-0858-11df-a26f-8b716316a5bc