Fix startup hangs when opening multiple files on Linux and Mac.
Seem there's either a critter in the wx2812 socket IPC code (the
socket processing was rewritten for the wx3 series) that could
cause Audacity to hang during startup if more than 1 file was
opened at the same time.  It was hit or miss, but more hit than
miss and was likely to occur when opening multiple files in a
file manager.

This fix was to stop using the wx2 IPC code and craft our own
basic startup messaging.



git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@13869 47890b92-0858-11df-a26f-8b716316a5bc